new object representation + print/display framework
