+++ h2o-py/h2o/expr.py
@@ -24,6 +24,7 @@ from h2o.utils.shared_utils import _is_fr, _py_tmp_key
 from h2o.model.model_base import ModelBase
 from h2o.expr_optimizer import optimize
 
+
 class ExprNode(object):
     """
     Composable Expressions: This module contains code for the lazy expression DAG.
@@ -258,8 +259,6 @@ class ExprNode(object):
         return h2o.api("POST /99/Rapids", data={"ast": expr, "session_id": h2o.connection().session_id})
 
 
-
-
 class ASTId:
     def __init__(self, name=None):
         if name is None:
@@ -268,8 +267,6 @@ class ASTId:
 
     def __repr__(self):
         return self.name
-
-
 
 
 class H2OCache(object):
@@ -398,7 +395,8 @@ class H2OCache(object):
 
     def _tabulate(self, tablefmt="simple", rollups=False, rows=10):
         """Pretty tabulated string of all the cached data, and column names"""
-        if not self.is_valid(): self.fill(rows=rows)
+        if not self.is_valid(): 
+            self.fill(rows=rows)
         # Pretty print cached data
         d = OrderedDict()
         # If also printing the rollup stats, build a full row-header
