+++ h2o-py/h2o/frame.py
@@ -21,7 +21,7 @@ import warnings
 
 import h2o
 from h2o.base import Keyed
-from h2o.display import H2ODisplay
+from h2o.display import H2ODisplay, H2OTableDisplay, display, in_ipy, in_zep, repr_def
 from h2o.exceptions import H2OTypeError, H2OValueError, H2ODeprecationWarning
 from h2o.expr import ExprNode
 from h2o.group_by import GroupBy
@@ -32,13 +32,14 @@ from h2o.utils.shared_utils import (_handle_numpy_array, _handle_pandas_data_fra
                                     _handle_python_lists, _is_list, _is_str_list, _py_tmp_key, _quoted,
                                     can_use_pandas, can_use_numpy, quote, normalize_slice, slice_is_normalized, 
                                     check_frame_id)
+from h2o.utils.threading import thread_context, thread_env
 from h2o.utils.typechecks import (assert_is_type, assert_satisfies, Enum, I, is_type, numeric, numpy_ndarray,
                                   numpy_datetime, pandas_dataframe, pandas_timestamp, scipy_sparse, U)
 
 __all__ = ("H2OFrame", )
 
 
-class H2OFrame(Keyed):
+class H2OFrame(Keyed, H2ODisplay):
     """
     Primary data store for H2O.
 
@@ -86,9 +87,9 @@ class H2OFrame(Keyed):
     __LOCAL_EXPANSION_ON_SINGLE_IMPORT__ = True
     __fdopen_kwargs = {} if PY2 else {'encoding': 'utf-8'}
 
-    #-------------------------------------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------------------------------------------------
     # Construction
-    #-------------------------------------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------------------------------------------------
 
     def __init__(self, python_obj=None, destination_frame=None, header=0, separator=",",
                  column_names=None, column_types=None, na_strings=None, skipped_columns=None):
@@ -119,7 +120,6 @@ class H2OFrame(Keyed):
         if cache is not None:
             fr._ex._cache.fill_from(cache)
         return fr
-
 
     def _upload_python_object(self, python_obj, destination_frame=None, header=0, separator=",",
                               column_names=None, column_types=None, na_strings=None, skipped_columns=None):
@@ -152,7 +152,6 @@ class H2OFrame(Keyed):
         tmp_file.close()  # close the streams
         self._upload_parse(tmp_path, destination_frame, 1, separator, column_names, column_types, na_strings, skipped_columns)
         os.remove(tmp_path)  # delete the tmp file
-
 
     def _upload_sparse_matrix(self, matrix, destination_frame=None):
         import scipy.sparse as sp
@@ -194,7 +193,6 @@ class H2OFrame(Keyed):
         H2OJob(h2o.api("POST /3/ParseSVMLight", data=p), "Parse").poll()
         self._ex._cache._id = destination_frame
         self._ex._cache.fill()
-
 
     @staticmethod
     def get_frame(frame_id, rows=10, rows_offset=0, cols=-1, full_cols=-1, cols_offset=0, light=False):
@@ -240,7 +238,6 @@ class H2OFrame(Keyed):
                 return fr
             return param
 
-
     def refresh(self):
         """
         Reload frame information from the backend H2O server.
@@ -265,11 +262,9 @@ class H2OFrame(Keyed):
         self._ex._cache.flush()
         self._frame(fill_cache=True)
 
-
-
-    #-------------------------------------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------------------------------------------------
     # Frame properties
-    #-------------------------------------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------------------------------------------------
 
     @property
     def key(self):
@@ -284,7 +279,6 @@ class H2OFrame(Keyed):
         >>> frame.key
         """
         return None if self._ex is None else self.frame_id
-
 
     @property
     def names(self):
@@ -307,7 +301,6 @@ class H2OFrame(Keyed):
     def names(self, value):
         self.set_names(value)
 
-
     @property
     def nrows(self):
         """
@@ -322,7 +315,6 @@ class H2OFrame(Keyed):
             self._ex._cache.flush()
             self._frame(fill_cache=True)
         return self._ex._cache.nrows
-
 
     @property
     def ncols(self):
@@ -341,7 +333,6 @@ class H2OFrame(Keyed):
             self._frame(fill_cache=True)
         return self._ex._cache.ncols
 
-
     @property
     def shape(self):
         """
@@ -356,7 +347,6 @@ class H2OFrame(Keyed):
         >>> iris.shape
         """
         return self.nrows, self.ncols
-
 
     @property
     def types(self):
@@ -424,7 +414,6 @@ class H2OFrame(Keyed):
             self._ex._cache._id = newid
             h2o.rapids("(rename \"{}\" \"{}\")".format(oldname, newid))
 
-
     def type(self, col):
         """
         The type for the given column.
@@ -452,7 +441,6 @@ class H2OFrame(Keyed):
                 return types[names[col]]
         raise H2OValueError("Column '%r' does not exist in the frame" % col)
 
-
     def _import_parse(self, path, pattern, destination_frame, header, separator, column_names, column_types, na_strings,
                       skipped_columns=None, custom_non_data_line_markers=None, partition_by=None, quotechar=None, escapechar=None):
         if H2OFrame.__LOCAL_EXPANSION_ON_SINGLE_IMPORT__ and is_type(path, str) and "://" not in path:  # fixme: delete those 2 lines, cf. PUBDEV-5717
@@ -462,7 +450,6 @@ class H2OFrame(Keyed):
                     skipped_columns, custom_non_data_line_markers, partition_by, quotechar, escapechar)
         return self
 
-
     def _upload_parse(self, path, destination_frame, header, sep, column_names, column_types, na_strings, skipped_columns=None,
                       quotechar=None, escapechar=None):
         ret = h2o.api("POST /3/PostFile", filename=path)
@@ -471,14 +458,12 @@ class H2OFrame(Keyed):
                     quotechar=quotechar, escapechar=escapechar)
         return self
 
-
     def _parse(self, rawkey, destination_frame="", header=None, separator=None, column_names=None, column_types=None,
                na_strings=None, skipped_columns=None, custom_non_data_line_markers=None, partition_by=None, quotechar=None,
                escapechar=None):
         setup = h2o.parse_setup(rawkey, destination_frame, header, separator, column_names, column_types, na_strings,
                                 skipped_columns, custom_non_data_line_markers, partition_by, quotechar, escapechar)
         return self._parse_raw(setup)
-
 
     def _parse_raw(self, setup):
         # Parse parameters (None values provided by setup)
@@ -512,7 +497,6 @@ class H2OFrame(Keyed):
         self._ex._cache._id = p["destination_frame"]
         self._ex._cache.fill()
 
-
     def filter_na_cols(self, frac=0.2):
         """
         Filter columns with proportion of NAs greater or equals than ``frac``.
@@ -531,7 +515,6 @@ class H2OFrame(Keyed):
         >>> include_cols2
         """
         return ExprNode("filterNACols", self, frac)._eager_scalar()
-
 
     def columns_by_type(self, coltype="numeric"):
         """
@@ -564,154 +547,8 @@ class H2OFrame(Keyed):
         assert_is_type(self, H2OFrame)
         return ExprNode("columnsByType", self, coltype)._eager_scalar()
 
-
     def __iter__(self):
         return (self[i] for i in range(self.ncol))
-
-    def __unicode__(self):
-        if sys.gettrace() is None:
-            if self._ex is None: return "This H2OFrame has been removed."
-            table = self._frame(fill_cache=True)._ex._cache._tabulate("simple", False)
-            nrows = "%d %s" % (self.nrow, "row" if self.nrow == 1 else "rows")
-            ncols = "%d %s" % (self.ncol, "column" if self.ncol == 1 else "columns")
-            return "%s\n\n[%s x %s]" % (table, nrows, ncols)
-        return ""
-
-    def __repr__(self):
-        if sys.gettrace() is None:
-            # PUBDEV-2278: using <method>? from IPython caused everything to dump
-            stk = traceback.extract_stack()
-            if not ("IPython" in stk[-2][0] and "info" == stk[-2][2]):
-                self.show()
-        return ""
-
-    def _has_content(self):
-        return self._ex and (self._ex._children or self._ex._cache._id)
-
-    def show(self, use_pandas=False, rows=10, cols=200):
-        """
-        Used by the H2OFrame.__repr__ method to print or display a snippet of the data frame.
-
-        If called from IPython, displays the results in HTML format. Otherwise, this prints a tabulated result.
-
-        :returns: snippet of the data frame.
-
-        :examples:
-
-        >>> from random import randrange
-        >>> import numpy as np
-        >>> row_num = randrange(1,10)
-        >>> col_num = randrange(1,10)
-        >>> python_lists = np.random.randint(-5,5, (row_num,col_num))
-        >>> h20frame = h2o.H2OFrame(python_obj=python_lists)
-        >>> h2oframe.show(use_pandas=False)
-        >>> h2oframe.show(use_pandas=True)
-        """
-        if self._ex is None:
-            print("This H2OFrame has been removed.")
-            return
-        if not self._has_content():
-            print("This H2OFrame is empty and not initialized.")
-            return
-        if self.nrows == 0:
-            print("This H2OFrame is empty.")
-            return
-        if not self._ex._cache.is_valid(): self._frame()._ex._cache.fill()
-        if H2ODisplay._in_zep():
-            print("%html " + self._ex._cache._tabulate("html", False, rows=rows))
-        elif H2ODisplay._in_ipy():
-            import IPython.display
-            if use_pandas and can_use_pandas():
-                IPython.display.display(self.head(rows=rows, cols=cols).as_data_frame(use_pandas=True))
-            else:
-                IPython.display.display_html(self._ex._cache._tabulate("html", False, rows=rows), raw=True)
-        else:
-            if use_pandas and can_use_pandas():
-                print(self.head(rows=rows, cols=cols).as_data_frame(use_pandas=True))
-            else:
-                s = self.__unicode__()
-                stk = traceback.extract_stack()
-                if "IPython" in stk[-3][0]:
-                    s = "\n%s" % s
-                try:
-                    print(s)
-                except UnicodeEncodeError:
-                    print(s.encode("ascii", "replace"))
-
-
-    def summary(self, return_data=False):
-        """
-        Display summary information about the frame.
-
-        Summary includes min/mean/max/sigma and other rollup data.
-
-        :param bool return_data: Return a dictionary of the summary output
-        :returns: Summary of information about the frame
-
-        :examples:
-
-        >>> frame = h2o.import_file("http://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris.csv")
-        >>> frame.summary()
-        """
-        if not self._has_content():
-            print("This H2OFrame is empty and not initialized.")
-            return self._ex._cache._data;
-        if not self._ex._cache.is_valid(): self._frame()._ex._cache.fill()
-        if not return_data:
-            if self.nrows == 0:
-                print("This H2OFrame is empty.")
-            elif H2ODisplay._in_zep():
-                print("%html " + self._ex._cache._tabulate("html", True))
-            elif H2ODisplay._in_ipy():
-                import IPython.display
-                IPython.display.display_html(self._ex._cache._tabulate("html", True), raw=True)
-            else:
-                print(self._ex._cache._tabulate("simple", True))
-        else:
-            return self._ex._cache._data
-
-
-    def describe(self, chunk_summary=False):
-        """
-        Generate an in-depth description of this H2OFrame.
-
-        This will print to the console the dimensions of the frame; names/types/summary statistics for each column;
-        and finally first ten rows of the frame.
-
-        :param bool chunk_summary: Retrieve the chunk summary along with the distribution summary
-        :returns: The dimensions of the frame; names/types/summary statistics for each column; first ten rows of the frame.
-
-        :examples:
-
-        >>> python_lists = [[1,2,3],[4,5,6],["a","b","c"],[1,0,1]]
-        >>> col_names=["num1","num2","str1","enum1"]
-        >>> dest_frame="newFrame"
-        >>> heads=-1
-        >>> sep=','
-        >>> col_types=['numeric','numeric','string','enum']
-        >>> na_str=['NA']
-        >>> h2oframe = h2o.H2OFrame(python_obj=python_lists,
-        ...                         destination_frame=dest_frame,
-        ...                         header=heads,
-        ...                         separator=sep,
-        ...                         column_names=col_names,
-        ...                         column_types=col_types,
-        ...                         na_strings=na_str)
-        >>> h2oframe.describe(chunk_summary=True)
-        """
-        if self._has_content():
-            res = h2o.api("GET /3/Frames/%s" % self.frame_id, data={"row_count": 10})["frames"][0]
-            self._ex._cache._fill_data(res)
-
-            print("Rows:{}".format(self.nrow))
-            print("Cols:{}".format(self.ncol))
-
-            # The chunk & distribution summaries are not cached, so must be pulled if chunk_summary=True.
-            if chunk_summary:
-                res["chunk_summary"].show()
-                res["distribution_summary"].show()
-            print("\n")
-        self.summary()
 
     def detach(self):
         """
@@ -730,13 +567,11 @@ class H2OFrame(Keyed):
         """
         self._ex = None
 
-
     def _frame(self, rows=10, rows_offset=0, cols=-1, cols_offset=0, fill_cache=False):
         self._ex._eager_frame()
         if fill_cache:
             self._ex._cache.fill(rows=rows, rows_offset=rows_offset, cols=cols, cols_offset=cols_offset)
         return self
-
 
     def head(self, rows=10, cols=200):
         """
@@ -763,7 +598,6 @@ class H2OFrame(Keyed):
         ncols = min(self.ncols, cols)
         newdt = self[:nrows, :ncols]
         return newdt._frame(rows=nrows, cols=cols, fill_cache=True)
-
 
     def tail(self, rows=10, cols=200):
         """
@@ -795,7 +629,6 @@ class H2OFrame(Keyed):
         newdt = self[start_idx:start_idx + nrows, :ncols]
         return newdt._frame(rows=nrows, cols=cols, fill_cache=True)
 
-
     def logical_negation(self):
         """
         Create a new H2OFrame equal to element-wise Logical NOT applied to the current frame.
@@ -809,7 +642,6 @@ class H2OFrame(Keyed):
         >>> frame.logical_negation()
         """
         return H2OFrame._expr(expr=ExprNode("not", self), cache=self._ex._cache)
-
 
     def _unop(self, op, rtype="real"):
         if self._is_frame:
@@ -927,7 +759,6 @@ class H2OFrame(Keyed):
     def __float__(self):
         return float(self.flatten())
 
-
     def flatten(self):
         """
         Convert a 1x1 frame into a scalar.
@@ -946,7 +777,6 @@ class H2OFrame(Keyed):
         """
         if self.shape != (1, 1): raise H2OValueError("Not a 1x1 Frame")
         return ExprNode("flatten", self)._eager_scalar()
-
 
     def getrow(self):
         """
@@ -979,7 +809,6 @@ class H2OFrame(Keyed):
             raise H2OValueError("This method can only be applied to single-row frames")
         return ExprNode("getrow", self)._eager_scalar()
 
-
     def mult(self, matrix):
         """
         Multiply this frame, viewed as a matrix, by another matrix.
@@ -998,7 +827,6 @@ class H2OFrame(Keyed):
             raise H2OValueError("Matrix is not compatible for multiplication with the current frame")
         return H2OFrame._expr(expr=ExprNode("x", self, matrix))
 
-
     def cos(self):
         """
         Create a new H2OFrame equal to element-wise cosine of the current frame.
@@ -1012,7 +840,6 @@ class H2OFrame(Keyed):
         >>> frame.cos()
         """
         return self._unop("cos")
-
 
     def sin(self):
         """
@@ -1028,7 +855,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("sin")
 
-
     def tan(self):
         """
         Create a new H2OFrame equal to element-wise tangent of the current frame.
@@ -1042,7 +868,6 @@ class H2OFrame(Keyed):
         >>> frame.tan()
         """
         return self._unop("tan")
-
 
     def acos(self):
         """
@@ -1058,7 +883,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("acos")
 
-
     def asin(self):
         """
         Create a new H2OFrame equal to element-wise arc sine of the current frame.
@@ -1072,7 +896,6 @@ class H2OFrame(Keyed):
         >>> frame.asin()
         """
         return self._unop("asin")
-
 
     def atan(self):
         """
@@ -1088,7 +911,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("atan")
 
-
     def cosh(self):
         """
         Create a new H2OFrame with values equal to the hyperbolic cosines of the values in the current frame.
@@ -1102,7 +924,6 @@ class H2OFrame(Keyed):
         >>> frame.cosh()
         """
         return self._unop("cosh")
-
 
     def sinh(self):
         """
@@ -1118,7 +939,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("sinh")
 
-
     def tanh(self):
         """
         Create a new H2OFrame equal to element-wise hyperbolic tangent of the current frame.
@@ -1132,7 +952,6 @@ class H2OFrame(Keyed):
         >>> frame.tanh()
         """
         return self._unop("tanh")
-
 
     def acosh(self):
         """
@@ -1148,7 +967,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("acosh")
 
-
     def asinh(self):
         """
         Create a new H2OFrame equal to element-wise inverse hyperbolic sine of the current frame.
@@ -1162,7 +980,6 @@ class H2OFrame(Keyed):
         >>> frame.asinh()
         """
         return self._unop("asinh")
-
 
     def atanh(self):
         """
@@ -1178,7 +995,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("atanh")
 
-
     def cospi(self):
         """
         Create a new H2OFrame equal to element-wise cosine of the current frame multiplied by Pi.
@@ -1192,7 +1008,6 @@ class H2OFrame(Keyed):
         >>> frame.cospi()
         """
         return self._unop("cospi")
-
 
     def sinpi(self):
         """
@@ -1208,7 +1023,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("sinpi")
 
-
     def tanpi(self):
         """
         Create a new H2OFrame equal to element-wise tangent of the current frame multiplied by Pi.
@@ -1222,7 +1036,6 @@ class H2OFrame(Keyed):
         >>> frame.tanpi()
         """
         return self._unop("tanpi")
-
 
     def abs(self):
         """
@@ -1238,7 +1051,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("abs")
 
-
     def sign(self):
         """
         Return new H2OFrame equal to signs of the values in the frame: -1 , +1, or 0.
@@ -1251,7 +1063,6 @@ class H2OFrame(Keyed):
         >>> iris.sign()
         """
         return self._unop("sign", rtype="int")
-
 
     def sqrt(self):
         """
@@ -1266,7 +1077,6 @@ class H2OFrame(Keyed):
         >>> frame.sqrt()
         """
         return self._unop("sqrt")
-
 
     def trunc(self):
         """
@@ -1292,7 +1102,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("trunc", rtype="int")
 
-
     def ceil(self):
         """
         Apply the ceiling function to the current frame.
@@ -1315,7 +1124,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("ceiling", rtype="int")
 
-
     def floor(self):
         """
         Apply the floor function to the current frame. ``floor(x)`` is the largest integer smaller or equal to ``x``.
@@ -1331,7 +1139,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("floor", rtype="int")
 
-
     def log(self):
         """
         Create a new H2OFrame equal to element-wise natural logarithm of the current frame.
@@ -1345,7 +1152,6 @@ class H2OFrame(Keyed):
         >>> frame.log()
         """
         return self._unop("log")
-
 
     def log10(self):
         """
@@ -1361,7 +1167,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("log10")
 
-
     def log1p(self):
         """
         Create a new H2Oframe equal to element-wise ``ln(1 + x)`` for each ``x`` in the current frame.
@@ -1375,7 +1180,6 @@ class H2OFrame(Keyed):
         >>> frame.log1p()
         """
         return self._unop("log1p")
-
 
     def log2(self):
         """
@@ -1391,7 +1195,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("log2")
 
-
     def exp(self):
         """
         Create a new H2OFrame equal to element-wise exponent (i.e. ``e^x``) of the current frame.
@@ -1405,7 +1208,6 @@ class H2OFrame(Keyed):
         >>> frame.exp()
         """
         return self._unop("exp")
-
 
     def expm1(self):
         """
@@ -1421,7 +1223,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("expm1")
 
-
     def gamma(self):
         """
         Create a new H2OFrame equal to element-wise gamma function of the current frame.
@@ -1435,7 +1236,6 @@ class H2OFrame(Keyed):
         >>> frame.gamma()
         """
         return self._unop("gamma")
-
 
     def lgamma(self):
         """
@@ -1451,7 +1251,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("lgamma")
 
-
     def digamma(self):
         """
         Create a new H2OFrame equal to element-wise digamma function of the current frame.
@@ -1466,7 +1265,6 @@ class H2OFrame(Keyed):
         """
         return self._unop("digamma")
 
-
     def trigamma(self):
         """
         Create a new H2OFrame equal to the element-wise trigamma function of the current frame.
@@ -1480,7 +1278,6 @@ class H2OFrame(Keyed):
         >>> frame.trigamma()
         """
         return self._unop("trigamma")
-
 
     @staticmethod
     def moment(year=None, month=None, day=None, hour=None, minute=None, second=None, msec=None, date=None, time=None):
@@ -1594,7 +1391,6 @@ class H2OFrame(Keyed):
         res._ex._cache._ncols = 1
         return res
 
-
     def unique(self, include_nas=False):
         """
         Extract the unique values in the column.
@@ -1610,7 +1406,6 @@ class H2OFrame(Keyed):
         >>> h2oframe.unique()
         """
         return H2OFrame._expr(expr=ExprNode("unique", self, include_nas))
-
 
     def levels(self):
         """
@@ -1632,7 +1427,6 @@ class H2OFrame(Keyed):
         lol = list(zip(*lol))
         return [[ll for ll in l if ll != ''] for l in lol]
 
-
     def nlevels(self):
         """
         Get the number of factor levels for each categorical column.
@@ -1648,7 +1442,6 @@ class H2OFrame(Keyed):
         """
         levels = self.levels()
         return [len(l) for l in levels] if levels else 0
-
 
     def set_level(self, level):
         """
@@ -1673,7 +1466,6 @@ class H2OFrame(Keyed):
         >>> newFrame
         """
         return H2OFrame._expr(expr=ExprNode("setLevel", self, level), cache=self._ex._cache)
-
 
     def set_levels(self, levels):
         """
@@ -1757,7 +1549,6 @@ class H2OFrame(Keyed):
 
         return self.set_names(new_names)
 
-
     def set_names(self, names):
         """
         Change names of all columns in the frame.
@@ -1780,7 +1571,6 @@ class H2OFrame(Keyed):
         assert_satisfies(names, len(names) == self.ncol)
         self._ex = ExprNode("colnames=", self, range(self.ncol), names)  # Update-in-place, but still lazy
         return self
-
 
     def set_name(self, col=None, name=None):
         """
@@ -1836,7 +1626,6 @@ class H2OFrame(Keyed):
             self._ex._cache._types[name] = self._ex._cache._types.pop(oldname)
         return
 
-
     def as_date(self, format):
         """
         Convert the frame (containing strings / categoricals) into the ``date`` format.
@@ -1853,7 +1642,6 @@ class H2OFrame(Keyed):
         if fr._ex._cache.types_valid():
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
-
 
     def cumsum(self,  axis=0):
         """
@@ -1872,7 +1660,6 @@ class H2OFrame(Keyed):
         >>> cumsum2 = foo[1].cumsum()
         """
         return H2OFrame._expr(expr=ExprNode("cumsum", self, axis), cache=self._ex._cache)
-
 
     def cumprod(self, axis=0):
         """
@@ -1893,7 +1680,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("cumprod", self, axis), cache=self._ex._cache)
 
-
     def cummin(self, axis=0):
         """
         Compute cumulative minimum over rows / columns of the frame.
@@ -1913,7 +1699,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("cummin", self, axis), cache=self._ex._cache)
 
-
     def cummax(self, axis=0):
         """
         Compute cumulative maximum over rows / columns of the frame.
@@ -1932,7 +1717,6 @@ class H2OFrame(Keyed):
         >>> cummax2
         """
         return H2OFrame._expr(expr=ExprNode("cummax", self, axis), cache=self._ex._cache)
-
 
     def prod(self, na_rm=False):
         """
@@ -1954,7 +1738,6 @@ class H2OFrame(Keyed):
         """
         return ExprNode("prod.na" if na_rm else "prod", self)._eager_scalar()
 
-
     def any(self):
         """
         Determine whether any element in the frame is either True, non-zero, or NA.
@@ -1968,7 +1751,6 @@ class H2OFrame(Keyed):
         >>> frame.any()
         """
         return bool(ExprNode("any", self)._eager_scalar())
-
 
     def any_na_rm(self):
         """
@@ -1984,7 +1766,6 @@ class H2OFrame(Keyed):
         """
         return bool(ExprNode("any.na", self)._eager_scalar())
 
-
     def all(self):
         """
         Determine whether every element in the frame is either True, non-zero, or NA.
@@ -1998,7 +1779,6 @@ class H2OFrame(Keyed):
         >>> frame.all()
         """
         return bool(ExprNode("all", self)._eager_scalar())
-
 
     def isnumeric(self):
         """
@@ -2017,7 +1797,6 @@ class H2OFrame(Keyed):
         # Return as False
         """
         return [bool(o) for o in ExprNode("is.numeric", self)._eager_scalar()]
-
 
     def isstring(self):
         """
@@ -2038,7 +1817,6 @@ class H2OFrame(Keyed):
         >>> newFrame.isstring()
         """
         return [bool(o) for o in ExprNode("is.character", self)._eager_scalar()]
-
 
     def isin(self, item):
         """
@@ -2061,7 +1839,6 @@ class H2OFrame(Keyed):
                 return functools.reduce(H2OFrame.__or__, (self == i for i in item))
         else:
             return self == item
-
 
     def kfold_column(self, n_folds=3, seed=-1):
         """
@@ -2086,7 +1863,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("kfold_column", self, n_folds, seed))._frame()  # want this to be eager!
 
-
     def modulo_kfold_column(self, n_folds=3):
         """
         Build a fold assignments column for cross-validation.
@@ -2107,7 +1883,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("modulo_kfold_column", self, n_folds))._frame()  # want this to be eager!
 
-
     def stratified_kfold_column(self, n_folds=3, seed=-1):
         """
         Build a fold assignment column with the constraint that each fold has the same class
@@ -2127,7 +1902,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(
             expr=ExprNode("stratified_kfold_column", self, n_folds, seed))._frame()  # want this to be eager!
-
 
     def structure(self):
         """
@@ -2484,7 +2258,6 @@ class H2OFrame(Keyed):
         if value_is_own_subframe:
             value._ex = None  # wipe out to keep ref counts correct
 
-
     def _is_frame_in_self(self, frame):
         if self._ex is frame._ex: return True
         if frame._ex._children is None: return False
@@ -2521,15 +2294,15 @@ class H2OFrame(Keyed):
         """
         if axis == 1:
             if not isinstance(index, list):
-                #If input is a string, i.e., "C1":
+                # If input is a string, i.e., "C1":
                 if is_type(index, str):
-                    #Check if index is an actual column(s) in the frame
+                    # Check if index is an actual column(s) in the frame
                     if index not in self.names:
                         raise H2OValueError("Column(s) selected to drop are not in original frame: %r" % index)
                     index = self.names.index(index)
-                #If input is an int indicating a column index, i.e., 3:
+                # If input is an int indicating a column index, i.e., 3:
                 elif is_type(index, int):
-                    #Check if index is an actual column index in the frame
+                    # Check if index is an actual column index in the frame
                     if index > self.ncol:
                         raise H2OValueError("Column index selected to drop is not part of the frame: %r" % index)
                     if index < 0:
@@ -2542,16 +2315,16 @@ class H2OFrame(Keyed):
                 return fr
 
             elif isinstance(index, list):
-                #If input is an int array indicating a column index, i.e., [3] or [1,2,3]:
+                # If input is an int array indicating a column index, i.e., [3] or [1,2,3]:
                 if is_type(index, [int]):
                     if max(index) > self.ncol:
                         raise H2OValueError("Column index selected to drop is not part of the frame: %r" % index)
                     if min(index) < 0:
                         raise H2OValueError("Column index selected to drop is not positive: %r" % index)
                     index = [-(i + 1) for i in index]
-                #If index is a string array, i.e., ["C1", "C2"]
+                # If index is a string array, i.e., ["C1", "C2"]
                 elif is_type(index, [str]):
-                    #Check if index is an actual column(s) in the frame
+                    # Check if index is an actual column(s) in the frame
                     if not set(index).issubset(self.names):
                         raise H2OValueError("Column(s) selected to drop are not in original frame: %r" % index)
                     index = [-(self.names.index(i) + 1) for i in index]
@@ -2568,7 +2341,7 @@ class H2OFrame(Keyed):
             return fr
         elif axis == 0:
             if is_type(index, [int]):
-                #Check if index is an actual column index in the frame
+                # Check if index is an actual column index in the frame
                 if max(index) > self.nrow:
                     raise H2OValueError("Row index selected to drop is not part of the frame: %r" % index)
                 if min(index) < 0:
@@ -2579,7 +2352,6 @@ class H2OFrame(Keyed):
             else:
                 raise ValueError("Invalid row indexes. Must be a list of int row indexes to drop from the H2OFrame.")
         return fr
-
 
     def pop(self, i):
         """
@@ -2609,7 +2381,6 @@ class H2OFrame(Keyed):
         col._ex._cache.ncols = 1
         col._ex._cache.names = [old_cache.names[i]]
         return col
-
 
     def quantile(self, prob=None, combine_method="interpolate", weights_column=None):
         """
@@ -2645,7 +2416,6 @@ class H2OFrame(Keyed):
                 return H2OFrame._expr(expr=ExprNode("quantile", merged, prob, combine_method, weights_column))
         return H2OFrame._expr(expr=ExprNode("quantile", self, prob, combine_method, weights_column))
 
-
     def concat(self, frames, axis=1):
         """
         Append multiple H2OFrames to this frame, column-wise or row-wise.
@@ -2673,7 +2443,6 @@ class H2OFrame(Keyed):
         else:
             df = self.rbind(frames)
         return df
-
 
     def cbind(self, data):
         """
@@ -2720,7 +2489,6 @@ class H2OFrame(Keyed):
             fr._ex._cache.types = None
         return fr
 
-
     def rbind(self, data):
         """
         Append data to this frame row-wise.
@@ -2756,7 +2524,6 @@ class H2OFrame(Keyed):
         fr = H2OFrame._expr(expr=ExprNode("rbind", self, *frames), cache=self._ex._cache)
         fr._ex._cache.nrows = self.nrow + sum(frame.nrow for frame in frames)
         return fr
-
 
     def split_frame(self, ratios=None, destination_frames=None, seed=None):
         """
@@ -2845,7 +2612,6 @@ class H2OFrame(Keyed):
         h2o.remove(tmp_runif)
         del tmp_runif
         return splits
-
 
     def group_by(self, by):
         """
@@ -2986,7 +2752,6 @@ class H2OFrame(Keyed):
             values = values2
         if group_by_frame is None: group_by_frame = "_"
 
-
         # This code below is needed to ensure the frame (self) exists on the server. Without it, self._ex._cache.fill()
         # fails with an assertion that ._id is None.
         # This code should be removed / reworked once we have a more consistent strategy of dealing with frames.
@@ -3002,7 +2767,6 @@ class H2OFrame(Keyed):
         self._ex._cache.flush()
         self._ex._cache.fill(10)
         return res
-
 
     def merge(self, other, all_x=False, all_y=False, by_x=None, by_y=None, method="auto"):
         """
@@ -3057,9 +2821,7 @@ class H2OFrame(Keyed):
         else:
             by_y = _getValidCols(by_y,other)
 
-
         return H2OFrame._expr(expr=ExprNode("merge", self, other, all_x, all_y, by_x, by_y, method))
-
 
     def relevel(self, y):
         """
@@ -3086,7 +2848,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("relevel", self, quote(y)))
 
-
     def relevel_by_frequency(self, weights_column=None, top_n=-1):
         """
         Reorder levels of all frame's factor columns according to their frequencies. The resulting frame
@@ -3098,7 +2859,6 @@ class H2OFrame(Keyed):
         :returns: A new Frame with the same shape and data, with levels re-ordered  
         """
         return H2OFrame._expr(expr=ExprNode("relevel.by.freq", self, weights_column, top_n))
-
 
     def insert_missing_values(self, fraction=0.1, seed=None):
         """
@@ -3133,7 +2893,6 @@ class H2OFrame(Keyed):
         self._ex._cache.flush()
         return self
 
-
     def min(self):
         """
         Show the minimum value of all frame entries.
@@ -3151,7 +2910,6 @@ class H2OFrame(Keyed):
         """
         return ExprNode("min", self)._eager_scalar()
 
-
     def max(self):
         """
         Show the maximum value of all frame entries.
@@ -3168,7 +2926,6 @@ class H2OFrame(Keyed):
         >>> iris["C4"].max()
         """
         return ExprNode("max", self)._eager_scalar()
-
 
     def sum(self, skipna=True, axis=0, **kwargs):
         """
@@ -3217,7 +2974,6 @@ class H2OFrame(Keyed):
         else:
             return ExprNode("sumNA" if skipna else "sum", self)._eager_scalar()
 
-
     def mean(self, skipna=True, axis=0, **kwargs):
         """
         Compute the frame's means by-column (or by-row).
@@ -3260,7 +3016,6 @@ class H2OFrame(Keyed):
         else:
             return new_frame.getrow()
 
-
     def skewness(self, na_rm=False):
         """
         Compute the skewness of each column in the frame.
@@ -3276,7 +3031,6 @@ class H2OFrame(Keyed):
         >>> h2oframe.skewness()
         """
         return ExprNode("skewness", self, na_rm)._eager_scalar()
-
 
     def kurtosis(self, na_rm=False):
         """
@@ -3297,7 +3051,6 @@ class H2OFrame(Keyed):
         """
         return ExprNode("kurtosis", self, na_rm)._eager_scalar()
 
-
     def nacnt(self):
         """
         Count of NAs for each column in this H2OFrame.
@@ -3310,7 +3063,6 @@ class H2OFrame(Keyed):
         >>> iris.nacnt()
         """
         return ExprNode("naCnt", self)._eager_scalar()
-
 
     def median(self, na_rm=False):
         """
@@ -3325,7 +3077,6 @@ class H2OFrame(Keyed):
         >>> iris.median()
         """
         return ExprNode("median", self, na_rm)._eager_scalar()
-
 
     def var(self, y=None, na_rm=False, use=None):
         """
@@ -3363,7 +3114,6 @@ class H2OFrame(Keyed):
             return ExprNode("var", self, y, use, symmetric)._eager_scalar()
         return H2OFrame._expr(expr=ExprNode("var", self, y, use, symmetric))._frame()
 
-
     def sd(self, na_rm=False):
         """
         Compute the standard deviation for each column in the frame.
@@ -3381,7 +3131,6 @@ class H2OFrame(Keyed):
         >>> newframe.sd()
         """
         return ExprNode("sd", self, na_rm)._eager_scalar()
-
 
     def cor(self, y=None, na_rm=False, use=None, method="Pearson"):
         """
@@ -3423,7 +3172,6 @@ class H2OFrame(Keyed):
         if use is None: use = "complete.obs" if na_rm else "everything"
         if self.nrow == 1 or (self.ncol == 1 and y.ncol == 1): return ExprNode("cor", self, y, use, method)._eager_scalar()
         return H2OFrame._expr(expr=ExprNode("cor", self, y, use, method))._frame()
-
 
     def distance(self, y, measure=None):
         """
@@ -3502,8 +3250,6 @@ class H2OFrame(Keyed):
         assert_is_type(compare_empty, bool)
         return H2OFrame._expr(expr=ExprNode("strDistance", self, y, measure, compare_empty))._frame()
 
-       
-
     def asfactor(self):
         """
         Convert column/columns in the current frame to categoricals.
@@ -3535,7 +3281,6 @@ class H2OFrame(Keyed):
         
         return fr
 
-
     def isfactor(self):
         """
         Test which columns in the current frame are categorical.
@@ -3554,7 +3299,6 @@ class H2OFrame(Keyed):
         """
         return [bool(o) for o in ExprNode("is.factor", self)._eager_scalar()]
 
-
     def anyfactor(self):
         """
         Determine if there are any categorical columns in the frame.
@@ -3568,7 +3312,6 @@ class H2OFrame(Keyed):
         >>> frame.anyfactor()
         """
         return bool(ExprNode("any.factor", self)._eager_scalar())
-
 
     def categories(self):
         """
@@ -3588,7 +3331,6 @@ class H2OFrame(Keyed):
             raise H2OValueError("Input is not a factor. This operation only applies to a single factor column")
         return self.levels()[0]
 
-
     def transpose(self):
         """
         Transpose rows and columns of this frame.
@@ -3606,7 +3348,6 @@ class H2OFrame(Keyed):
         >>> h2oframe.transpose()
         """
         return H2OFrame._expr(expr=ExprNode("t", self))
-
 
     def strsplit(self, pattern):
         """
@@ -3685,7 +3426,6 @@ class H2OFrame(Keyed):
         fr._ex._cache.ncols = self.ncol
         return fr
 
-
     def trim(self):
         """
         Trim white space on the left and right of strings in a single-column H2OFrame.
@@ -3705,7 +3445,6 @@ class H2OFrame(Keyed):
         fr._ex._cache.nrows = self.nrow
         fr._ex._cache.ncol = self.ncol
         return fr
-
 
     def substring(self, start_index, end_index=None):
         """
@@ -3728,7 +3467,6 @@ class H2OFrame(Keyed):
         fr._ex._cache.nrows = self.nrow
         fr._ex._cache.ncol = self.ncol
         return fr
-
 
     def lstrip(self, set=" "):
         """
@@ -3753,7 +3491,6 @@ class H2OFrame(Keyed):
         fr._ex._cache.nrows = self.nrow
         fr._ex._cache.ncol = self.ncol
         return fr
-
 
     def rstrip(self, set=" "):
         """
@@ -3781,7 +3518,6 @@ class H2OFrame(Keyed):
         fr._ex._cache.ncol = self.ncol
         return fr
 
-
     def entropy(self):
         """
         For each string compute its Shannon entropy, if the string is empty the entropy is 0.
@@ -3797,7 +3533,6 @@ class H2OFrame(Keyed):
         fr._ex._cache.nrows = self.nrow
         fr._ex._cache.ncol = self.ncol
         return fr
-
 
     def num_valid_substrings(self, path_to_words):
         """
@@ -3824,7 +3559,6 @@ class H2OFrame(Keyed):
         fr._ex._cache.ncol = self.ncol
         return fr
 
-
     def nchar(self):
         """
         Count the length of each string in a single-column H2OFrame of string type.
@@ -3837,7 +3571,6 @@ class H2OFrame(Keyed):
         >>> iris[4].nchar()
         """
         return H2OFrame._expr(expr=ExprNode("strlen", self))
-
 
     def table(self, data2=None, dense=True):
         """
@@ -3855,7 +3588,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("table", self, data2, dense)) if data2 is not None else H2OFrame._expr(
             expr=ExprNode("table", self, dense))
-
 
     def hist(self, breaks="sturges", plot=True, save_plot_path=None, **kwargs):
         """
@@ -3929,7 +3661,6 @@ class H2OFrame(Keyed):
         else:
             hist["density"] = hist["counts"] / (hist["breaks"].difflag1() * hist["counts"].sum())
             return decorate_plot_result(res=hist)
-
 
     def isax(self, num_words, max_cardinality, optimize_card=False, **kwargs):
         """
@@ -4355,7 +4086,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("replacefirst", self, pattern, replacement, ignore_case))
 
-
     def gsub(self, pattern, replacement, ignore_case=False):
         """
         Globally substitute occurrences of pattern in a string with replacement.
@@ -4374,7 +4104,6 @@ class H2OFrame(Keyed):
         >>> iris["C5"].gsub("s","z",ignore_case=False)
         """
         return H2OFrame._expr(expr=ExprNode("replaceall", self, pattern, replacement, ignore_case))
-
 
     def interaction(self, factors, pairwise, max_factors, min_occurrence, destination_frame=None):
         """
@@ -4412,7 +4141,6 @@ class H2OFrame(Keyed):
         return h2o.interaction(data=self, factors=factors, pairwise=pairwise, max_factors=max_factors,
                                min_occurrence=min_occurrence, destination_frame=destination_frame)
 
-
     def toupper(self):
         """
         Translate characters from lower to upper case for a particular column.
@@ -4427,7 +4155,7 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("toupper", self), cache=self._ex._cache)
 
-    def grep(self,pattern, ignore_case = False, invert = False, output_logical = False):
+    def grep(self, pattern, ignore_case=False, invert=False, output_logical=False):
         """
         Searches for matches to argument `pattern` within each element
         of a string column.
@@ -4464,7 +4192,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("tolower", self), cache=self._ex._cache)
 
-
     def rep_len(self, length_out):
         """
         Create a new frame replicating the current frame.
@@ -4495,7 +4222,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("rep_len", self, length_out))
 
-
     def scale(self, center=True, scale=True):
         """
         Center and/or scale the columns of the current frame.
@@ -4517,7 +4243,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("scale", self, center, scale), cache=self._ex._cache)
 
-
     def signif(self, digits=6):
         """
         Round doubles/floats to the given number of significant digits.
@@ -4534,7 +4259,6 @@ class H2OFrame(Keyed):
         >>> h2o_data.signif(digits = 2)
         """
         return H2OFrame._expr(expr=ExprNode("signif", self, digits), cache=self._ex._cache)
-
 
     def round(self, digits=0):
         """
@@ -4554,7 +4278,6 @@ class H2OFrame(Keyed):
         >>> h2o_data.round(digits = 0)
         """
         return H2OFrame._expr(expr=ExprNode("round", self, digits), cache=self._ex._cache)
-
 
     def asnumeric(self):
         """
@@ -4576,7 +4299,6 @@ class H2OFrame(Keyed):
             fr._ex._cache.types = {k: "real" for k in fr._ex._cache.types.keys()}
         return fr
 
-
     def ascharacter(self):
         """
         Convert all columns in the frame into strings.
@@ -4594,7 +4316,6 @@ class H2OFrame(Keyed):
             fr._ex._cache.types = {k: "string" for k in fr._ex._cache.types.keys()}
         return fr
 
-
     def na_omit(self):
         """
         Remove rows with NAs from the H2OFrame.
@@ -4611,7 +4332,6 @@ class H2OFrame(Keyed):
         fr = H2OFrame._expr(expr=ExprNode("na.omit", self), cache=self._ex._cache)
         fr._ex._cache.nrows = -1
         return fr
-
 
     def difflag1(self):
         """
@@ -4639,7 +4359,6 @@ class H2OFrame(Keyed):
         fr = H2OFrame._expr(expr=ExprNode("difflag1", self), cache=self._ex._cache)
         return fr
 
-
     def isna(self):
         """
         For each element in an H2OFrame, determine if it is NA or not.
@@ -4666,7 +4385,6 @@ class H2OFrame(Keyed):
             fr._ex._cache.types = {"isNA(%s)" % n: "int" for n in self._ex._cache.names}
         return fr
 
-
     def year(self):
         """
         Extract the "year" part from a date column.
@@ -4690,7 +4408,6 @@ class H2OFrame(Keyed):
         if fr._ex._cache.types_valid():
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
-
 
     def month(self):
         """
@@ -4716,7 +4433,6 @@ class H2OFrame(Keyed):
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
 
-
     def week(self):
         """
         Extract the "week" part from a date column.
@@ -4740,7 +4456,6 @@ class H2OFrame(Keyed):
         if fr._ex._cache.types_valid():
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
-
 
     def day(self):
         """
@@ -4766,7 +4481,6 @@ class H2OFrame(Keyed):
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
 
-
     def dayOfWeek(self):
         """
         Extract the "day-of-week" part from a date column.
@@ -4790,7 +4504,6 @@ class H2OFrame(Keyed):
         if fr._ex._cache.types_valid():
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
-
 
     def hour(self):
         """
@@ -4816,7 +4529,6 @@ class H2OFrame(Keyed):
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
 
-
     def minute(self):
         """
         Extract the "minute" part from a date column.
@@ -4840,7 +4552,6 @@ class H2OFrame(Keyed):
         if fr._ex._cache.types_valid():
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
-
 
     def second(self):
         """
@@ -4866,7 +4577,6 @@ class H2OFrame(Keyed):
             fr._ex._cache.types = {k: "int" for k in self._ex._cache.types.keys()}
         return fr
 
-
     def runif(self, seed=None):
         """
         Generate a column of random numbers drawn from a uniform distribution [0,1) and
@@ -4888,7 +4598,6 @@ class H2OFrame(Keyed):
         fr._ex._cache.nrows = self.nrow
         return fr
 
-
     def stratified_split(self, test_frac=0.2, seed=-1):
         """
         Construct a column that can be used to perform a random stratified split.
@@ -4906,7 +4615,6 @@ class H2OFrame(Keyed):
         >>> h2oframe[1].stratified_split(test_frac=0.2, seed=-1)
         """
         return H2OFrame._expr(expr=ExprNode('h2o.random_stratified_split', self, test_frac, seed))
-
 
     def match(self, table, nomatch=0):
         """
@@ -4928,7 +4636,6 @@ class H2OFrame(Keyed):
         >>> matchFrame
         """
         return H2OFrame._expr(expr=ExprNode("match", self, table, nomatch, None))
-
 
     def cut(self, breaks, labels=None, include_lowest=False, right=True, dig_lab=3):
         """
@@ -4966,7 +4673,6 @@ class H2OFrame(Keyed):
                             cache=self._ex._cache)
         fr._ex._cache.types = {k: "enum" for k in self.names}
         return fr
-
 
     def which(self):
         """
@@ -5033,7 +4739,6 @@ class H2OFrame(Keyed):
         """
         return H2OFrame._expr(expr=ExprNode("which.min", self, skipna, axis))
 
-
     def ifelse(self, yes, no):
         """
         Equivalent to ``[y if t else n for t,y,n in zip(self,yes,no)]``.
@@ -5058,7 +4763,6 @@ class H2OFrame(Keyed):
         >>> newFrame
         """
         return H2OFrame._expr(expr=ExprNode("ifelse", self, yes, no))
-
 
     def apply(self, fun=None, axis=0):
         """
@@ -5085,10 +4789,9 @@ class H2OFrame(Keyed):
         res = lambda_to_expr(fun)
         return H2OFrame._expr(expr=ExprNode("apply", self, 1 + (axis == 0), *res))
 
-
-    #-------------------------------------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------------------------------------------------
     # Synonyms + Deprecated
-    #-------------------------------------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------------------------------------------------
     # Here we have all methods that are provided as alternative names to some other names defined above. This also
     # includes methods that we rename as part of the deprecation process (but keeping the old name for the sake of
     # backward compatibility). We gather them all down here to have a slightly cleaner code.
@@ -5188,15 +4891,6 @@ class H2OFrame(Keyed):
         """
         return [self.nrow, self.ncol]
 
-    #@property
-    #def frame_id(self):
-    #    """Same as ``frame.id``."""
-    #    return self.id
-
-    #@frame_id.setter
-    #def frame_id(self, value):
-    #    self.id = value
-
     @staticmethod
     def from_python(python_obj, destination_frame=None, header=0, separator=",", column_names=None,
                     column_types=None, na_strings=None):
@@ -5204,16 +4898,202 @@ class H2OFrame(Keyed):
         return H2OFrame(python_obj, destination_frame, header, separator, column_names, column_types,
                         na_strings)
 
-
     def ischaracter(self):
         """[DEPRECATED] Use ``frame.isstring()``."""
         return self.isstring()
+    
+    # -------------------------------
+    # H2OFrame representation methods
+    # -------------------------------
+    
+    def _str_edge_cases(self):
+        if self._ex is None:
+            return "H2OFrame has been removed."
+        if not self._has_content():
+            return "H2OFrame is empty and not initialized."
+        if self.nrows == 0:
+            return "H2OFrame is empty."
+        # ensure that the cached frame has data to display
+        rows = thread_env('rows', 10)
+        cols = thread_env('cols', -1)
+        self._frame()._ex._cache.fill(rows=rows, cols=cols)
+        return None
+
+    def _repr_(self):
+        # no need to pollute debug string with cell values, headers should be enough
+        return repr_def(self, attributes=['_ex'])
+    
+    def _to_str(self, format_=None):
+        fr_str = self._str_edge_cases()
+        if fr_str is not None:
+            return fr_str
+
+        rows = thread_env('rows', 10)
+        cols = thread_env('cols', 200)
+        use_pandas = thread_env('pandas', H2OTableDisplay.prefer_pandas())
+        if use_pandas:
+            df = self.head(rows=rows, cols=cols).as_data_frame(use_pandas=True)
+            table = df.to_html() if format_ == 'html' else df.to_string()
+        else:
+            tablefmt = dict(
+                pretty='simple',
+                html='html'
+            ).get(format_, 'plain')
+            table = self._frame(fill_cache=True)._ex._cache._tabulate(tablefmt, rollups=False, rows=rows)
+            
+        nrows = "{nrow} row{s}".format(nrow=self.nrow, s="" if self.nrow == 1 else "s")
+        ncols = "{ncol} column{s}".format(ncol=self.ncol, s="" if self.ncol == 1 else "s")
+        template = dict(
+            pretty="{table}\n\n[{nrows} x {ncols}]",
+            html="{table}<br><pre>[{nrows} x {ncols}]</pre>"
+        ).get(format_, "{table}\n[{nrows} x {ncols}])")
+        return template.format(table=table, nrows=nrows, ncols=ncols)
+
+    def _str_(self):
+        return self._to_str()
+    
+    def _str_pretty_(self):
+        return self._to_str('pretty')
+
+    def _str_html_(self):
+        return self._to_str('html')
+
+    def _has_content(self):
+        return self._ex and (self._ex._children or self._ex._cache._id)
+
+    def show(self, use_pandas=None, rows=10, cols=200):
+        # changing default behaviour, suggesting to use pandas by default as soon as it's available.
+        if use_pandas is None:
+            use_pandas = can_use_pandas()
+        with thread_context(rows=rows, cols=cols, pandas=use_pandas):
+            display(self)
+
+    def show_old(self, use_pandas=False, rows=10, cols=200):
+        """
+        Print a snippet of the data frame.
+        :examples:
+
+        >>> from random import randrange
+        >>> import numpy as np
+        >>> row_num = randrange(1,10)
+        >>> col_num = randrange(1,10)
+        >>> python_lists = np.random.randint(-5,5, (row_num,col_num))
+        >>> h20frame = h2o.H2OFrame(python_obj=python_lists)
+        >>> h2oframe.show(use_pandas=False)
+        >>> h2oframe.show(use_pandas=True)
+        """
+        
+        if self._ex is None:
+            print("This H2OFrame has been removed.")
+            return
+        if not self._has_content():
+            print("This H2OFrame is empty and not initialized.")
+            return
+        if self.nrows == 0:
+            print("This H2OFrame is empty.")
+            return
+        
+        if not self._ex._cache.is_valid():
+            self._frame()._ex._cache.fill()
+            
+        if in_zep():
+            print("%html " + self._ex._cache._tabulate("html", False, rows=rows))
+        elif in_ipy():
+            import IPython.display
+            if use_pandas and can_use_pandas():
+                IPython.display.display(self.head(rows=rows, cols=cols).as_data_frame(use_pandas=True))
+            else:
+                IPython.display.display_html(self._ex._cache._tabulate("html", False, rows=rows), raw=True)
+        else:
+            if use_pandas and can_use_pandas():
+                print(self.head(rows=rows, cols=cols).as_data_frame(use_pandas=True))
+            else:
+                s = self.__unicode__()
+                stk = traceback.extract_stack()
+                if "IPython" in stk[-3][0]:
+                    s = "\n%s" % s
+                try:
+                    print(s)
+                except UnicodeEncodeError:
+                    print(s.encode("ascii", "replace"))
+
+    def summary(self, return_data=False):
+        """
+        Display summary information about the frame.
+
+        Summary includes min/mean/max/sigma and other rollup data.
+
+        :param bool return_data: Return a dictionary of the summary output
+        :returns: Summary of information about the frame
+
+        :examples:
+
+        >>> frame = h2o.import_file("http://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris.csv")
+        >>> frame.summary()
+        """
+        if not self._has_content():
+            print("This H2OFrame is empty and not initialized.")
+            return self._ex._cache._data;
+        if not self._ex._cache.is_valid(): self._frame()._ex._cache.fill()
+        if not return_data:
+            if self.nrows == 0:
+                print("This H2OFrame is empty.")
+            elif in_zep():
+                print("%html " + self._ex._cache._tabulate("html", True))
+            elif in_ipy():
+                import IPython.display
+                IPython.display.display_html(self._ex._cache._tabulate("html", True), raw=True)
+            else:
+                print(self._ex._cache._tabulate("simple", True))
+        else:
+            return self._ex._cache._data
+
+    def describe(self, chunk_summary=False):
+        """
+        Generate an in-depth description of this H2OFrame.
+
+        This will print to the console the dimensions of the frame; names/types/summary statistics for each column;
+        and finally first ten rows of the frame.
+
+        :param bool chunk_summary: Retrieve the chunk summary along with the distribution summary
+        :returns: The dimensions of the frame; names/types/summary statistics for each column; first ten rows of the frame.
+
+        :examples:
+
+        >>> python_lists = [[1,2,3],[4,5,6],["a","b","c"],[1,0,1]]
+        >>> col_names=["num1","num2","str1","enum1"]
+        >>> dest_frame="newFrame"
+        >>> heads=-1
+        >>> sep=','
+        >>> col_types=['numeric','numeric','string','enum']
+        >>> na_str=['NA']
+        >>> h2oframe = h2o.H2OFrame(python_obj=python_lists,
+        ...                         destination_frame=dest_frame,
+        ...                         header=heads,
+        ...                         separator=sep,
+        ...                         column_names=col_names,
+        ...                         column_types=col_types,
+        ...                         na_strings=na_str)
+        >>> h2oframe.describe(chunk_summary=True)
+        """
+        if self._has_content():
+            res = h2o.api("GET /3/Frames/%s" % self.frame_id, data={"row_count": 10})["frames"][0]
+            self._ex._cache._fill_data(res)
+
+            print("Rows:{}".format(self.nrow))
+            print("Cols:{}".format(self.ncol))
+
+            # The chunk & distribution summaries are not cached, so must be pulled if chunk_summary=True.
+            if chunk_summary:
+                res["chunk_summary"].show()
+                res["distribution_summary"].show()
+            print("\n")
+        self.summary()
 
 
-
-#-----------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------
 # Helpers
-#-----------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------
 
 def _getValidCols(by_idx, fr):  # so user can input names of the columns as well is idx num
     tmp = []
@@ -5256,9 +5136,6 @@ def _binop(lhs, op, rhs, rtype=None):
     if rtype is not None and res._ex._cache._names is not None:
         res._ex._cache._types = {name: rtype for name in res._ex._cache._names}
     return res
-
-
-
 
 def generatePandaEnumCols(pandaFtrain, cname, nrows, domainL):
     """
