+++ h2o-py/h2o/model/__init__.py
@@ -1,27 +1,24 @@
-from .autoencoder import H2OAutoEncoderModel
-from .binomial import H2OBinomialModel
-from .binomial_uplift import H2OBinomialUpliftModel
-from .clustering import H2OClusteringModel
+import sys
+
+from h2o.utils.mixin import register_submodule
+
 from .confusion_matrix import ConfusionMatrix
-from .dim_reduction import H2ODimReductionModel
 from .metrics_base import MetricsBase
 from .model_base import ModelBase
-from .model_future import H2OModelFuture
-from .multinomial import H2OMultinomialModel
-from .ordinal import H2OOrdinalModel
-from .regression import H2ORegressionModel
+from .models import *
 from .segment_models import H2OSegmentModels
-from .metrics_base import H2OAutoEncoderModelMetrics
-from .metrics_base import H2OBinomialModelMetrics
-from .metrics_base import H2OClusteringModelMetrics
-from .metrics_base import H2ODimReductionModelMetrics
-from .metrics_base import H2OMultinomialModelMetrics
-from .metrics_base import H2OOrdinalModelMetrics
-from .metrics_base import H2ORegressionModelMetrics
-from .metrics_base import H2OBinomialUpliftModelMetrics
 
 # order here impacts order of presentation in generated documentation
 __all__ = ["ModelBase", "MetricsBase", 
            "H2OBinomialModel", "H2OMultinomialModel", "H2ORegressionModel", "H2OOrdinalModel",
-           "H2OClusteringModel", "H2ODimReductionModel", "H2OAutoEncoderModel",
-           "ConfusionMatrix", "H2OModelFuture", "H2OSegmentModels", "H2OBinomialUpliftModel"]
+           "H2OClusteringModel", "H2ODimReductionModel", "H2OAutoEncoderModel", "H2OBinomialUpliftModel",
+           "ConfusionMatrix",  "H2OSegmentModels", ]
+
+
+# for full backwards compatibility after having moved some modules to `h2o.model.models` submodule.
+# not that users don't need to import those submodules in client 
+# code except for some (old) top functions in `regression` submodule.
+module = sys.modules[__name__]
+for mod in ['regression']:
+    register_submodule(module, name=mod, module=sys.modules["h2o.model.models.%s" % mod])
+
