+++ h2o-py/h2o/model/metrics/generic.py
@@ -1,0 +1,7 @@
+from __future__ import division, absolute_import, print_function, unicode_literals
+
+from h2o.model import MetricsBase
+
+
+class H2ODefaultModelMetrics(MetricsBase):
+    pass
