+++ h2o-py/h2o/model/model_future.py
@@ -1,3 +1,4 @@
+# DELETE ME I'm useless !!!
 
 class H2OModelFuture(object):
     """
