+++ h2o-py/h2o/model/models/__init__.py
@@ -1,0 +1,13 @@
+from .anomaly_detection import H2OAnomalyDetectionModel
+from .autoencoder import H2OAutoEncoderModel
+from .binomial import H2OBinomialModel
+from .clustering import H2OClusteringModel
+from .coxph import H2OCoxPHModel, H2OCoxPHMojoModel
+from .dim_reduction import H2ODimReductionModel
+from .multinomial import H2OMultinomialModel
+from .ordinal import H2OOrdinalModel
+from .regression import H2ORegressionModel
+from .uplift import H2OBinomialUpliftModel
+from .word_embedding import H2OWordEmbeddingModel
+
+__all__ = [s for s in dir() if not s.startswith('_')]
