+++ h2o-py/h2o/model/models/anomaly_detection.py
@@ -1,8 +1,8 @@
 # -*- encoding: utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+from h2o.model import ModelBase
 from h2o.utils.shared_utils import can_use_pandas
-from .model_base import ModelBase
 
 
 class H2OAnomalyDetectionModel(ModelBase):
