+++ h2o-py/h2o/model/models/coxph.py
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from .model_base import ModelBase
+from h2o.model import ModelBase
 
 
 class H2OCoxPHModel(ModelBase):
@@ -17,10 +17,6 @@ class H2OCoxPHModel(ModelBase):
     def coefficients_table(self):
         """Coefficients table."""
         return self._model_json["output"]["coefficients_table"]
-
-    def show_summary(self):
-        print(self._summary())
-
 
     def _summary(self):
         """Prints summary information about this model."""
