+++ h2o-py/h2o/model/models/ordinal.py
@@ -2,19 +2,15 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import h2o
+from h2o.exceptions import H2OValueError
+from h2o.frame import H2OFrame
+from h2o.model import ModelBase
+from h2o.model.extensions import has_extension
 from h2o.utils.compatibility import *  # NOQA
 from h2o.utils.typechecks import assert_is_type
-from .extensions import has_extension
-from .model_base import ModelBase
-from ..exceptions import H2OValueError
-from ..frame import H2OFrame
 
 
 class H2OOrdinalModel(ModelBase):
-
-    def _make_model(self):
-        return H2OOrdinalModel()
-
 
     def confusion_matrix(self, data):
         """
@@ -25,7 +21,6 @@ class H2OOrdinalModel(ModelBase):
         assert_is_type(data, H2OFrame)
         j = h2o.api("POST /3/Predictions/models/%s/frames/%s" % (self._id, data.frame_id))
         return j["model_metrics"][0]["cm"]["table"]
-
 
     def hit_ratio_table(self, train=False, valid=False, xval=False):
         """
@@ -45,7 +40,6 @@ class H2OOrdinalModel(ModelBase):
         for k, v in zip(list(tm.keys()), list(tm.values())): m[k] = None if v is None else v.hit_ratio_table()
         return list(m.values())[0] if len(m) == 1 else m
 
-
     def mean_per_class_error(self, train=False, valid=False, xval=False):
         """
         Retrieve the mean per class error across all classes
@@ -64,7 +58,6 @@ class H2OOrdinalModel(ModelBase):
         m = {}
         for k, v in zip(list(tm.keys()), list(tm.values())): m[k] = None if v is None else v.mean_per_class_error()
         return list(m.values())[0] if len(m) == 1 else m
-
 
     def plot(self, timestep="AUTO", metric="AUTO", save_plot_path=None, **kwargs):
         """
