+++ h2o-py/h2o/plot/__init__.py
@@ -1,5 +1,4 @@
-from ._matplotlib import get_matplotlib_pyplot
-from ._polycollection import get_polycollection
-from ._plot_result import decorate_plot_result, is_decorated_plot_result, RAISE_ON_FIGURE_ACCESS
+from ._matplotlib import *
+from ._plot_result import *
 
-__all__ = ["decorate_plot_result", "get_matplotlib_pyplot", "is_decorated_plot_result", "get_polycollection", "RAISE_ON_FIGURE_ACCESS"]
+__all__ = [s for s in dir() if not s.startswith('_')]
