+++ h2o-py/h2o/plot/_matplotlib.py
@@ -27,3 +27,35 @@ def get_matplotlib_pyplot(server, raise_if_not_available=False):
             raise e
         print("`matplotlib` library is required for this function!")
         return None
+
+
+def get_polycollection(server, raise_if_not_available=False):
+    try:
+        from matplotlib.collections import PolyCollection as polycoll
+        return polycoll
+    except ImportError as e:
+        if raise_if_not_available:
+            raise e
+        print("`matplotlib` library is required for this function!")
+        return None
+    
+    
+def get_matplotlib_cm(function_name):
+    try:
+        from matplotlib import cm
+        return cm
+    except ImportError:
+        print('matplotlib library is required for 3D plots for function {0}'.format(function_name))
+        return None
+
+
+def get_mplot3d_axes(function_name):
+    try:
+        # noinspection PyUnresolvedReferences
+        from mpl_toolkits.mplot3d import Axes3D
+        return Axes3D
+    except ImportError:
+        print("`mpl_toolkits.mplot3d` library is required for function {0}!".format(function_name))
+        return None
+
+
