+++ h2o-py/h2o/utils/typechecks.py
@@ -130,8 +130,6 @@ else:
     _primitive_type = (str, int, float, bool, bytes)
 
 
-
-
 def is_type(var, *args):
     """
     Return True if the variable is of the specified type(s) and False otherwise.
@@ -142,9 +140,9 @@ def is_type(var, *args):
     return _check_type(var, U(*args))
 
 
-#-----------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------
 # Special types
-#-----------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------
 
 class MagicType(object):
     """Abstract "special" type."""
