+++ h2o-py/tests/testdir_algos/isofor/pyunit_isofor_valid.py
@@ -4,7 +4,8 @@ sys.path.insert(1, os.path.join("..","..",".."))
 import h2o
 from tests import pyunit_utils
 from h2o.estimators.isolation_forest import H2OIsolationForestEstimator
-from h2o.model.metrics_base import H2OAnomalyDetectionModelMetrics, H2OBinomialModelMetrics
+from h2o.model import H2OBinomialModelMetrics
+from h2o.model.metrics.anomaly_detection import H2OAnomalyDetectionModelMetrics
 from pandas.testing import assert_frame_equal
